Complete adding updating and deleting order

+++ MyShop/Model/Order_Product.cs
@@ -12,10 +12,10 @@ namespace MyShop.Model
         public int OrderId   { get; set; }
         public int ProductId { get; set; }
 
-        public decimal Price { get; set; }
+        public int Price { get; set; }
         public int Amount { get; set; }
 
-        public Order_Product(int orderProductId, int orderId, int productId, decimal price, int amount)
+        public Order_Product(int orderProductId, int orderId, int productId, int price, int amount)
         {
             OrderId = orderId;
             ProductId = productId;
